+++ src/components/CustomAttachmentCard.vue
