+++ src/views/course/edit/CourseEditSectionView.vue
