+++ src/views/course/edit/CourseTaskView.vue
