+++ src/views/teacher/TeacherCourseView.vue
@@ -1,12 +1,8 @@
 <script>
-import router from '@/router';
-import base_course_card from '../base_module/base_course_card.vue'
+import cource_card from '../widgets/course-card.vue'
 export default {
-  components: { base_course_card },
-  methods:{
-    cource_detail(){
-      router.push('/course')
-    }
+  components: {
+    cource_card
   }
 }
 </script>
@@ -24,13 +20,13 @@ export default {
     <div id="main_cource">
       <h4>使用中</h4>
       <div id="use_card">
-        <base_course_card @click="cource_detail"/>
-        <base_course_card />
+        <cource_card />
+        <cource_card />
       </div>
       <h4>待发布</h4>
       <div id="wait_card">
-        <base_course_card />
-        <base_course_card />
+        <cource_card />
+        <cource_card />
       </div>
     </div>
   </div>
