+++ src/views/teacher/TeacherMainView.vue
@@ -1,8 +1,11 @@
 <script>
-import base_message from '../base_module/base_message.vue'
-import base_student_card from '../base_module/base_student_card.vue'
+import one_stu from '../widgets/one_stu.vue'
+import message from '../widgets/message.vue'
 export default {
-  components: { base_message, base_student_card },
+  components: {
+    one_stu,
+    message
+  }
 }
 </script>
 <template>
@@ -22,7 +25,7 @@ export default {
       <v-virtual-scroll :items="['1']" max-height="620">
         <div id="stu_detail">
           <template v-for="item in 40" :key="item">
-            <base_student_card />
+            <one_stu />
           </template>
         </div>
       </v-virtual-scroll>
@@ -35,7 +38,7 @@ export default {
       <div id="message">
         <v-infinite-scroll height="528" mode="manual" @load="jump" load-more-text="查看更多消息">
           <template v-for="item in 20" :key="item">
-            <base_message />
+            <message />
           </template>
         </v-infinite-scroll>
         <div id="jumpmsg_column">
